requirements.txt
# Requirements for Inception-Style RAG Debate System

python-dotenv>=1.0.0

llama-index>=0.10.0
llama-index-llms-groq>=0.1.0
llama-index-embeddings-huggingface>=0.1.0

gradio>=4.0.0
sentence-transformers>=2.2.0